+++ src/ood/weatherProject/ForecastDisplay.java
@@ -1,7 +1,7 @@
 package ood.weatherProject;
 
 public class ForecastDisplay {
-    public void update(float temperature,float humidity,float pressure){
+    public void update(double temperature,double humidity,double pressure){
         //update program ...
         System.out.println("Updating forecast");
         System.out.printf("***Temperature:%.2f***Humidity:%.2f***Pressure:%.2f***\n", temperature, humidity, pressure);
